Refactor ToDo List
